fix: rendering bugs
