+++ app/games/minesweeper/config.ts
@@ -1,0 +1,17 @@
+export const basicConfig = {
+  beginner: {
+    width: 9,
+    height: 9,
+    mines: 10,
+  },
+  intermediate: {
+    width: 16,
+    height: 16,
+    mines: 40,
+  },
+  expert: {
+    width: 16,
+    height: 30,
+    mines: 99,
+  },
+};
