+++ app/games/minesweeper/minesweeper.module.css
@@ -14,6 +14,7 @@
   border-right: 5px solid #838383;
   border-bottom: 5px solid #838383;
   width: 2rem;
+  height: 2.1rem;
   padding: 0.25rem 0.5rem;
   background-color: #bfbfbf;
   cursor: pointer;
@@ -46,4 +47,5 @@
   border: 1px solid black;
   display: flex;
   justify-content: space-between;
+  padding: 0.25rem;
 }
